Fix Bounds & implement loading triggers
